working on issue 8
